KE-7: 리팩토링: 카카오맵API 파일로 분리
