+++ src/pages/trade/BookTradeDetail.jsx
@@ -2,6 +2,7 @@ import React, {useEffect, useRef, useState} from "react";
 import { useNavigate } from 'react-router-dom';
 import Button from "../../components/ui/Button";
 import { deleteTrade, getTrade } from "../../services/trade/tradeAPI";
+import kakaoMap from "../../services/kakao/kakaoMap";
 
 export default function TradeDetails() {
   const tradeId = 3;
@@ -26,39 +27,8 @@ export default function TradeDetails() {
   };
 
   useEffect(() => {
-    const kakaoMaps = window.kakao.maps;
-
     const handleResize = () => {
-      if (trade && kakaoMaps) {
-        const container = document.getElementById('map');
-        const options = {
-          center: new kakaoMaps.LatLng(trade.latitude, trade.longitude),
-          level: 3
-        };
-        const map = new kakaoMaps.Map(container, options);
-        mapRef.current = map;
-
-        const marker = new kakaoMaps.Marker({
-          position: options.center
-        });
-        const geocoder = new kakaoMaps.services.Geocoder();
-
-        const updateAddressInfo = () => {
-          const mapInstance = mapRef.current;
-          if (mapInstance) {
-            geocoder.coord2Address(trade.longitude, trade.latitude, (result, status) => {
-              if (status === kakaoMaps.services.Status.OK) {
-                const address = result[0]?.road_address?.address_name || '장소를 표시할 수 없습니다.';
-                setAddressInfo(address);
-              }
-            });
-          }
-        };
-        updateAddressInfo();
-        map.setDraggable(false);
-        map.setZoomable(false);
-        marker.setMap(map);
-      }
+      kakaoMap(trade, mapRef, setAddressInfo);
     };
 
     handleResize();
